Update download message duration in BDayCard
Changed the message displayed to users regarding the image download
time from "Please allow a few seconds for the image to download!"
to "Please allow up to 2 minutes for the image to download!"
to better set expectations for users experiencing longer download
times. This adjustment aims to improve user experience and reduce
potential confusion.